Run unittests.
